+++ IPlug/IGraphics/IKeyboardControl.h
@@ -90,7 +90,7 @@ been declared, so it is propbably best to include it in your plug-in's main
 class IKeyboardControl: public IControl
 {
 public:
-  IKeyboardControl(IPlugBase& plug, int x, int y, int minNote, int nOctaves, IBitmap& regularKeys, IBitmap& sharpKey, const int *pKeyCoords = 0):
+  IKeyboardControl(IPlugBase& plug, float x, float y, int minNote, int nOctaves, IBitmap& regularKeys, IBitmap& sharpKey, const int *pKeyCoords = 0):
     IControl(plug, IRECT(x, y, regularKeys), kNoParameter),
     mMinNote(minNote), mNumOctaves(nOctaves), mRegularKeys(regularKeys), mSharpKey(sharpKey),
     mOctaveWidth(pRegularKeys->W * 7), mMaxKey(nOctaves * 12), mKey(-1)
